refactor(subway): modify the code structure
- [입출력] 인터페이스를 통한 코드 구조 수정
