--- src/main/java/subway/service/FeatureInterface.java
@@ -1,9 +1,0 @@
-package subway.service;
-
-import java.util.Scanner;
-
-public interface FeatureInterface {
-    boolean isInput(String input);
-    void chooseFeature(String input);
-    void chooseSubwayFeature(String input, Scanner scanner);
-}
