+++ src/main/java/subway/service/StationService.java
@@ -2,12 +2,14 @@ package subway.service;
 
 import subway.domain.Station;
 import subway.repository.StationRepository;
+import subway.service.abstraction.feature.FeatureInterface;
+import subway.type.InputType;
 import subway.type.StationType;
 import subway.view.OutputView;
 
 import java.util.Scanner;
 
-public class StationService {
+public class StationService implements FeatureInterface {
     public static void initializeStations() {
         StationRepository.addStation(new Station(StationType.EDUCATION_UNIVERSITY.getStation()));
         StationRepository.addStation(new Station(StationType.GANGNAM.getStation()));
@@ -19,17 +21,37 @@ public class StationService {
     }
 
     public static void manageStation(Scanner scanner) {
-        FeatureService featureService = new FeatureService();
+        InputService inputService = new InputService();
+        StationService stationService = new StationService();
 
         System.out.println();
         while (true) {
             OutputView.printStationManagementScreen();
             String stationInput = scanner.next();
-            if (featureService.isInput(stationInput)) {
-                featureService.chooseFeature(stationInput);
+            if (inputService.isInput(stationInput)) {
+                stationService.chooseFeature(stationInput);
                 break;
             }
             OutputView.printInvalidFeatureChoiceException();
+        }
+    }
+
+    @Override
+    public void chooseFeature(String input) {
+        if (input.equals(InputType.INPUT_ONE.getInput())) {
+            // TODO: 역 등록 기능 구현
+            return;
+        }
+        if (input.equals(InputType.INPUT_TWO.getInput())) {
+            // TODO: 역 삭제 기능 구현
+            return;
+        }
+        if (input.equals(InputType.INPUT_THREE.getInput())) {
+            // TODO: 역 조회 기능 구현
+            return;
+        }
+        if (input.equals(InputType.INPUT_BACK.getInput())) {
+            return;
         }
     }
 }
