+++ src/main/java/subway/service/abstraction/feature/FeatureInterface.java
@@ -1,0 +1,5 @@
+package subway.service.abstraction.feature;
+
+public interface FeatureInterface {
+    void chooseFeature(String input);
+}
