+++ src/main/java/subway/service/abstraction/feature/SubwayFeatureInterface.java
@@ -1,0 +1,7 @@
+package subway.service.abstraction.feature;
+
+import java.util.Scanner;
+
+public interface SubwayFeatureInterface {
+    void chooseSubwayFeature(String input, Scanner scanner);
+}
