+++ src/main/java/subway/service/abstraction/input/InputInterface.java
@@ -1,0 +1,5 @@
+package subway.service.abstraction.input;
+
+public interface InputInterface {
+    boolean isInput(String input);
+}
